TL;DR: speed up chcon for up to 1.5x to 2x with Go 1.16+.

1. pkg/pwalkdir: add
    
    This is a copy-paste implementation of pwalk but utilizing
    filepath.WalkDir instead (added in Go 1.16). As WalkDir do not call
    stat(2) on every entry, it is much (2x on my machine) faster than Walk
    (see below).
    
    The idea is to have pwalk and pwalkdir coexist for some time, then
    gradually retire pwalk once Go < 1.16 is unsupported.

2. Chcon: speedup for Go 1.16+
    
    Using the newly added pwalkdir, the benchmark added by and described
    in commit ba956cafe09661021bf shows a considerable speed improvement
    (0.8s with pwalk, 0.5s with pwalkdir, 0.9s with chcon from GNU coreutils).