@kolyshkin 2021-08-06
A few non-scientific benchmarks (testdata is Linux kernel source tree, find is for reference only)

| Op | Time |
| --- | --- |
| find -print >/dev/null | 0.19s |
| chcon -R | 0.92s |
| Before this PR (pwalk) | 0.83s |
| This PR (pwalkdir) + go 1.16 | 0.51s |
| This PR (pwalkdir) + go 1.17rc2 | 0.45s |