@thaJeztah 2021-08-08
Looks like you forgot to remove some references that used the alias;

```
  level=error msg="Running error: buildir: analysis skipped: errors in package: [/home/runner/work/selinux/selinux/pkg/pwalkdir/pwalkdir_test.go:124:42: undeclared name: WalkFunc /home/runner/work/selinux/selinux/pkg/pwalkdir/pwalkdir_test.go:127:34: undeclared name: WalkFunc /home/runner/work/selinux/selinux/pkg/pwalkdir/pwalkdir_test.go:141:8: undeclared name: WalkFunc /home/runner/work/selinux/selinux/pkg/pwalkdir/pwalkdir_test.go:153:24: cannot use filepath.WalkDir (value of type func(root string, fn fs.WalkDirFunc) error) as walkerFunc value in struct literal /home/runner/work/selinux/selinux/pkg/pwalkdir/pwalkdir_test.go:154:21: cannot use Walk (value of type func(root string, walkFn fs.WalkDirFunc) error) as walkerFunc value in struct literal]"
```